+++ frontend/src/lib/constants/routes.constants.ts
@@ -19,7 +19,13 @@ export const ROUTE_ID_GROUP_NNS = "/(nns)";
 export const ROUTE_ID_GROUP_LIST = "/(list)";
 export const ROUTE_ID_GROUP_DETAILS = "/(detail)";
 
-export const ROUTE_ID_GROUPS = [ROUTE_ID_GROUP_APP, ROUTE_ID_GROUP_UNIVERSE, ROUTE_ID_GROUP_NNS, ROUTE_ID_GROUP_LIST, ROUTE_ID_GROUP_DETAILS];
+export const ROUTE_ID_GROUPS = [
+  ROUTE_ID_GROUP_APP,
+  ROUTE_ID_GROUP_UNIVERSE,
+  ROUTE_ID_GROUP_NNS,
+  ROUTE_ID_GROUP_LIST,
+  ROUTE_ID_GROUP_DETAILS,
+];
 
 export const UNIVERSE_PARAM = "u";
